!152 S10P12A810-321 [feat] getCommunityList 구현 * [fix] like, scrap 수정 - 로그인 하지 않는 사용자의 경우 canlike, canscrap false로 설정
* [fix] ReadCommunity 에러 수정
- fetch join으로 인해 tag의 개수만큼 comment가 생기는 에러 수정
- api uri v0에서 v1으로 수정

* [refactor] 기존 for문을 stream 코드로 refactor

* [chore] swagger 메세지, TODO 삭제

* [fix] token이 없는 요청에서 상세 조회가 안되는 에러 수정

* [feat] getCommunityList 구현
- 최적화 반드시 필요

+++ backend/src/main/java/com/f17coders/classhub/module/domain/community/dto/response/CommunityListRes.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Builder
 public record CommunityListRes(
         List<CommunityListDetailRes> communityList,
-        int totalPages
+        long totalPages
 ) {
 }
